The following cases in setAsync were not handled correctly:
Case 1:
A) T1 gets a cas version for an existing key.
B) T2 deletes the key.
C) T1 tries to set the key with the cas version but fails with KEY_NOT_FOUND status.

Case 2:
A) T1 finds that the key is missing.
B) T2 adds the key.
C) T1 tries to add the key but fails with ITEM_NOT_STORED.